fix(ci): split oversized coverage checker tests
